sistema_medico/requirements.txt
django>=4.0
djangorestframework